project1/project1/requirements.txt
dj-database-url==0.4.2
django-crispy-forms==1.7.0
gunicorn==19.7.1
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.2
